+++ src/test/DependencyContainer.spec.ts
@@ -25,14 +25,14 @@ describe('DependencyContainer', () => {
 		it('Should return same value everytime', () => {
 			// Arrange
 			const VALUE = 'abc';
-			let container: IDependencyContainer = new DependencyContainer(),
+			const container: IDependencyContainer = new DependencyContainer(),
 				internalContainer: Container = container['_container'];
 			
 			// Act
 			container.bindConstant<string>(IDENTIFIER, VALUE);
 
 			// Assert
-			let instance_1st = internalContainer.get<string>(IDENTIFIER),
+			const instance_1st = internalContainer.get<string>(IDENTIFIER),
 				instance_2nd = internalContainer.get<string>(IDENTIFIER),
 				instance_3rd = internalContainer.get<string>(IDENTIFIER);
 
@@ -50,7 +50,7 @@ describe('DependencyContainer', () => {
 			// Arrange
 			const VALUE_OLD = 'abc',
 				VALUE_NEW = 'xyz';
-			let container: IDependencyContainer = new DependencyContainer(),
+			const container: IDependencyContainer = new DependencyContainer(),
 				internalContainer: Container = container['_container'];
 			
 			// Act
@@ -58,7 +58,7 @@ describe('DependencyContainer', () => {
 			container.bindConstant<string>(IDENTIFIER, VALUE_NEW);
 
 			// Assert
-			let instance_1st = internalContainer.get<string>(IDENTIFIER),
+			const instance_1st = internalContainer.get<string>(IDENTIFIER),
 				instance_2nd = internalContainer.get<string>(IDENTIFIER),
 				instance_3rd = internalContainer.get<string>(IDENTIFIER);
 
@@ -76,9 +76,9 @@ describe('DependencyContainer', () => {
 	describe('bind', () => {
 		it('Should register dependency to internal container, with string identifier', () => {
 			// Arrange
-			let container: IDependencyContainer = new DependencyContainer(),
-				internalContainer: Container = container['_container'],
-				resolveInstance: IDummy;
+			const container: IDependencyContainer = new DependencyContainer(),
+				internalContainer: Container = container['_container'];
+			let resolveInstance: IDummy;
 			
 			// Act
 			container.bind<IDummy>('abc', Dummy); // String identifier
@@ -91,9 +91,9 @@ describe('DependencyContainer', () => {
 		
 		it('Should register dependency to internal container, with symbol identifier', () => {
 			// Arrange
-			let container: IDependencyContainer = new DependencyContainer(),
-				internalContainer: Container = container['_container'],
-				resolveInstance: IDummy;
+			const container: IDependencyContainer = new DependencyContainer(),
+				internalContainer: Container = container['_container'];
+			let resolveInstance: IDummy;
 			
 			// Act
 			container.bind<IDummy>(IDENTIFIER, Dummy); // Symbol identifier
@@ -106,14 +106,14 @@ describe('DependencyContainer', () => {
 		
 		it('Should return same instance everytime, when registering as singleton', () => {
 			// Arrange
-			let container: IDependencyContainer = new DependencyContainer(),
+			const container: IDependencyContainer = new DependencyContainer(),
 				internalContainer: Container = container['_container'];
 			
 			// Act
 			container.bind<IDummy>(IDENTIFIER, Dummy).asSingleton();
 
 			// Assert
-			let instance_1st = internalContainer.get<IDummy>(IDENTIFIER),
+			const instance_1st = internalContainer.get<IDummy>(IDENTIFIER),
 				instance_2nd = internalContainer.get<IDummy>(IDENTIFIER),
 				instance_3rd = internalContainer.get<IDummy>(IDENTIFIER);
 
@@ -132,14 +132,14 @@ describe('DependencyContainer', () => {
 		
 		it('Should create new instance everytime, when registering as transient', () => {
 			// Arrange
-			let container: IDependencyContainer = new DependencyContainer(),
+			const container: IDependencyContainer = new DependencyContainer(),
 				internalContainer: Container = container['_container'];
 			
 			// Act
 			container.bind<IDummy>(IDENTIFIER, Dummy).asTransient(); // Default behavior
 
 			// Assert
-			let instance_1st = internalContainer.get<IDummy>(IDENTIFIER),
+			const instance_1st = internalContainer.get<IDummy>(IDENTIFIER),
 				instance_2nd = internalContainer.get<IDummy>(IDENTIFIER),
 				instance_3rd = internalContainer.get<IDummy>(IDENTIFIER);
 
@@ -160,7 +160,7 @@ describe('DependencyContainer', () => {
 	describe('dispose', () => {
 		it('Should throw exception if called after disposal', () => {
 			// Arrange
-			let container: IDependencyContainer = new DependencyContainer();
+			const container: IDependencyContainer = new DependencyContainer();
 			container.bind<IDummy>(IDENTIFIER, Dummy);
 			
 			// Act
@@ -182,9 +182,9 @@ describe('DependencyContainer', () => {
 	describe('resolve', () => {
 		it('Should get dependency from internal container', () => {
 			// Arrange
-			let container: IDependencyContainer = new DependencyContainer(),
-				internalContainer: Container = container['_container'],
-				resolveInstance: IDummy;
+			const container: IDependencyContainer = new DependencyContainer(),
+				internalContainer: Container = container['_container'];
+			let resolveInstance: IDummy;
 			
 			// Act
 			internalContainer.bind<IDummy>(IDENTIFIER).to(Dummy);
@@ -197,8 +197,8 @@ describe('DependencyContainer', () => {
 		
 		it('Should return null if no dependency is found', () => {
 			// Arrange
-			let container: IDependencyContainer = new DependencyContainer(),
-				resolveInstance: IDummy;
+			const container: IDependencyContainer = new DependencyContainer();
+			let resolveInstance: IDummy;
 			
 			// Act
 			resolveInstance = container.resolve<IDummy>(IDENTIFIER);
@@ -211,7 +211,7 @@ describe('DependencyContainer', () => {
 	describe('isBound', () => {
 		it('Should check if a dependency is bound or not.', () => {
 			// Arrange
-			let container: IDependencyContainer = new DependencyContainer(),
+			const container: IDependencyContainer = new DependencyContainer(),
 				internalContainer: Container = container['_container'];
 			
 			// Act
@@ -229,7 +229,7 @@ describe('DependencyContainer', () => {
 	describe('unbind', () => {
 		it('Should not resolve unbound dependency anymore.', () => {
 			// Arrange
-			let container: IDependencyContainer = new DependencyContainer(),
+			const container: IDependencyContainer = new DependencyContainer(),
 				internalContainer: Container = container['_container'];
 			
 			internalContainer.bind<IDummy>(IDENTIFIER).to(Dummy);